OtherInfoWindow rename to OtherInfoView, las clases que se usban en moredetails.fulllogic movidas al nuevo paquete view para mantener funcionalidad (excepto OtherInfoView esa si corresponde)
