+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/external/artist/artists/ArtistExternalInjector.kt
@@ -1,6 +1,7 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.artists
 import ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.LastFMToArtistResolver
 import ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.LastFMToArtistResolverImpl
+import ayds.lisboa.songinfo.moredetails.fulllogic.view.LastFMAPI
 import retrofit2.Retrofit
 import retrofit2.converter.scalars.ScalarsConverterFactory
 
@@ -9,7 +10,7 @@ object ArtistExternalInjector {
     private const val RETROFIT_URL = "https://ws.audioscrobbler.com/2.0/"
 
     private val retrofit = createRetrofit()
-    private val lastFMAPI = retrofit.create(ayds.lisboa.songinfo.moredetails.fulllogic.LastFMAPI::class.java)
+    private val lastFMAPI = retrofit.create(LastFMAPI::class.java)
 
 
     private fun createRetrofit(): Retrofit {
