+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/view/DataBase.kt
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic
+package ayds.lisboa.songinfo.moredetails.fulllogic.view
 
 import android.content.ContentValues
 import android.content.Context
